+++ biosimulations/angular.json
@@ -1434,28 +1434,27 @@
         }
       }
     },
-    "messages-dispatch": {
-      "root": "libs/messages/dispatch",
-      "sourceRoot": "libs/messages/dispatch/src",
+    "messages": {
+      "root": "libs/messages",
+      "sourceRoot": "libs/messages/src",
       "projectType": "library",
       "schematics": {},
       "architect": {
         "lint": {
-          "builder": "@nrwl/linter:lint",
+          "builder": "@angular-eslint/builder:lint",
           "options": {
-            "linter": "eslint",
             "tsConfig": [
-              "libs/messages/dispatch/tsconfig.lib.json",
-              "libs/messages/dispatch/tsconfig.spec.json"
+              "libs/messages/tsconfig.lib.json",
+              "libs/messages/tsconfig.spec.json"
             ],
-            "exclude": ["**/node_modules/**", "!libs/messages/dispatch/**/*"]
+            "exclude": ["**/node_modules/**", "!libs/messages/**/*"]
           }
         },
         "test": {
           "builder": "@nrwl/jest:jest",
           "options": {
-            "jestConfig": "libs/messages/dispatch/jest.config.js",
-            "tsConfig": "libs/messages/dispatch/tsconfig.spec.json",
+            "jestConfig": "libs/messages/jest.config.js",
+            "tsConfig": "libs/messages/tsconfig.spec.json",
             "passWithNoTests": true
           }
         }
