+++ biosimulations/apps/dispatch-service/src/app/app.controller.spec.ts
@@ -10,7 +10,6 @@ import {
   Transport,
   NatsOptions,
 } from '@nestjs/microservices';
-import { exception } from 'console';
 
 describe('AppController', () => {
   let app: TestingModule;
