+++ biosimulations/libs/messages/README.md
@@ -1,0 +1,7 @@
+# messages
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `ng test messages` to execute the unit tests via [Jest](https://jestjs.io).
