--- biosimulations/libs/messages/dispatch/README.md
@@ -1,7 +1,0 @@
-# messages-dispatch
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `ng test messages-dispatch` to execute the unit tests via [Jest](https://jestjs.io).
